biencoder/requirements.txt
# torch
# pysolr==3.8.1
emoji==0.5.3
regex==2019.8.19
# matplotlib==3.1.0
tqdm==4.32.1
nltk==3.4.5
numpy==1.17.2
segtok==1.5.7
flair==0.6.1
pytorch-transformers==1.2.0
transformers==3.1.0
colorama==0.4.3
termcolor==1.1.0
faiss-cpu>=1.6.1
# custom
# gensim==3.8.3
pandas
# numpy
# jupyter
# textdistance
# beautifulsoup4
# SPARQLWrapper
# seaborn
# matplotlib
# sklearn
# scikit-learn-extra
# tabulate
# addict
uvicorn
fastapi
# supervisor
gatenlp
